refactoring modal
